+++ app/src/main/java/com/moroom/android/data/model/CheckedTextViewData.java
@@ -1,4 +1,4 @@
-package com.moroom.android;
+package com.moroom.android.data.model;
 
 public class CheckedTextViewData {
